Possible spelling: Relaxed.sum/product.
