+++ Sources/ComplexModule/Complex+AlgebraicField.swift
@@ -123,4 +123,17 @@ extension Complex: AlgebraicField {
     }
     return nil
   }
+  
+  @inlinable
+  public static func _relaxedAdd(_ a: Self, _ b: Self) -> Self {
+    Complex(Relaxed.sum(a.x, b.x), Relaxed.sum(a.y, b.y))
+  }
+  
+  @inlinable
+  public static func _relaxedMul(_ a: Self, _ b: Self) -> Self {
+    Complex(
+      Relaxed.sum(Relaxed.product(a.x, b.x), -Relaxed.product(a.y, b.y)),
+      Relaxed.sum(Relaxed.product(a.x, b.y),  Relaxed.product(a.y, b.x))
+    )
+  }
 }
