+++ Sources/RealModule/CMakeLists.txt
@@ -17,7 +17,8 @@ add_library(RealModule
   Float16+Real.swift
   Float80+Real.swift
   Real.swift
-  RealFunctions.swift)
+  RealFunctions.swift
+  RelaxedArithmetic.swift)
 set_target_properties(RealModule PROPERTIES
   INTERFACE_INCLUDE_DIRECTORIES ${CMAKE_Swift_MODULE_DIRECTORY})
 target_link_libraries(RealModule PUBLIC
